+++ governance/hardhat.config.js
@@ -12,6 +12,13 @@ const {
   parseForkUrl,
 } = require('@unlock-protocol/hardhat-helpers')
 
+// zksync
+if (process.env.ZK_SYNC) {
+  require('@matterlabs/hardhat-zksync-solc')
+  require('@matterlabs/hardhat-zksync-verify')
+  require('@matterlabs/hardhat-zksync-upgradable')
+}
+
 const settings = {
   optimizer: {
     enabled: true,
@@ -32,6 +39,16 @@ if (process.env.RUN_FORK) {
   copySync('.openzeppelin/mainnet.json', '.openzeppelin/unknown-31337.json')
 }
 
+// add tenderly if needed
+if (process.env.TENDERLY_FORK) {
+  const tdly = require('@tenderly/hardhat-tenderly')
+  tdly.setup()
+  networks.tenderly = {
+    url: process.env.TENDERLY_FORK,
+    accounts: networks.mainnet.accounts,
+  }
+}
+
 // tasks
 require('./tasks/balance')
 require('./tasks/safe')
@@ -40,14 +57,27 @@ require('./tasks/verify')
 require('./tasks/deploy')
 require('./tasks/set')
 require('./tasks/unlock')
+require('./tasks/lock')
 
 /**
  * @type import('hardhat/config').HardhatUserConfig
  */
-module.exports = {
+const config = {
   networks,
   etherscan,
   solidity: {
     compilers: [{ version: '0.8.21', settings }],
   },
+  sourcify: {
+    enabled: true,
+  },
 }
+
+if (process.env.ZK_SYNC) {
+  config.zksolc = {
+    version: 'latest',
+    settings: {},
+  }
+}
+
+module.exports = config