+++ governance/scripts/deployments/publicLock.js
@@ -1,35 +1,46 @@
-const { ethers, run } = require('hardhat')
-const { isLocalhost } = require('@unlock-protocol/hardhat-helpers')
-const contracts = require('@unlock-protocol/contracts')
+const { ethers } = require('hardhat')
+const {
+  ADDRESS_ZERO,
+  deployContract,
+  copyAndBuildContractsAtVersion,
+} = require('@unlock-protocol/hardhat-helpers')
 
 async function main({ publicLockVersion }) {
   // fetch chain info
   const [signer] = await ethers.getSigners()
 
-  let PublicLock
-  if (publicLockVersion) {
-    const { abi, bytecode } = contracts[`PublicLockV${publicLockVersion}`]
-    console.log(
-      `PUBLIC LOCK > Deploying lock template for released version ${publicLockVersion} with signer ${signer.address}`
-    )
-    PublicLock = await ethers.getContractFactory(abi, bytecode)
-  } else {
+  if (!publicLockVersion) {
     throw Error('Need to set --public-lock-version')
   }
 
-  const publicLock = await PublicLock.deploy()
-  await publicLock.waitForDeployment()
-  const { hash } = await publicLock.deploymentTransaction()
-  const publicLockAddress = await publicLock.getAddress()
-
-  // eslint-disable-next-line no-console
   console.log(
-    `PUBLIC LOCK > deployed v${await publicLock.publicLockVersion()} to : ${publicLockAddress} (tx: ${hash})`
+    `PUBLIC LOCK > Deploying lock template for released version ${publicLockVersion} with signer ${signer.address}`
   )
 
-  if (!isLocalhost()) {
-    await run('verify:verify', { address: publicLockAddress })
-  }
+  const [qualifiedPath] = await copyAndBuildContractsAtVersion(__dirname, [
+    { contractName: 'PublicLock', version: publicLockVersion },
+  ])
+
+  const { contract: publicLock, address: publicLockAddress } =
+    await deployContract(qualifiedPath)
+
+  console.log(
+    `PUBLIC LOCK > deployed v${await publicLock.publicLockVersion()} to : ${publicLockAddress}`
+  )
+  // initialize the template to prevent someone else from doing it
+  const { hash: txInitHash } = await publicLock.initialize(
+    signer.address,
+    0,
+    ADDRESS_ZERO,
+    0,
+    0,
+    ''
+  )
+  console.log(`PUBLIC LOCK > Template initialized (tx: ${txInitHash})`)
+
+  // renounce the manager role that was added during initilization
+  const { hash: txRenounceHash } = await publicLock.renounceLockManager()
+  console.log(`PUBLIC LOCK > manager role revoked (tx: ${txRenounceHash})`)
 
   return publicLockAddress
 }