+++ smart-contracts/test/Unlock/migrateLockData/v10.js
@@ -74,9 +74,8 @@ describe('upgradeLock / data migration v9 > v10', () => {
     const [unlockOwner, creator] = await ethers.getSigners()
 
     // deploy latest implementation
-    ;[PublicLockPast, PublicLockLatest] = await copyAndBuildContractsAtVersion(
-      dirname,
-      [
+    const [pathPublicLockPast, pathPublicLockLatest] =
+      await copyAndBuildContractsAtVersion(dirname, [
         {
           contractName: 'PublicLock',
           version: previousVersionNumber,
@@ -85,8 +84,11 @@ describe('upgradeLock / data migration v9 > v10', () => {
           contractName: 'PublicLock',
           version: previousVersionNumber + 1,
         },
-      ]
-    )
+      ])
+
+    PublicLockPast = await ethers.getContractFactory(pathPublicLockPast)
+    PublicLockLatest = await ethers.getContractFactory(pathPublicLockLatest)
+
     // deploy latest version
     const publicLockLatest = await PublicLockLatest.deploy()
     await publicLockLatest.deployed()
@@ -115,7 +117,7 @@ describe('upgradeLock / data migration v9 > v10', () => {
     const args = [
       60 * 60 * 24 * 30, // 30 days
       ADDRESS_ZERO,
-      ethers.utils.parseEther('0.01'),
+      ethers.utils.parseEther('0.01').toString(),
       1000, // available keys
       'A neat upgradeable lock!',
     ]