+++ unlock-app/src/components/interface/checkout/main/Metadata.tsx
@@ -169,7 +169,7 @@ export const MetadataInputs = ({
           }}
           render={({ field: { onChange, ref, onBlur } }) => {
             return (
-              <div className="grid gap-1.5">
+              <div className="grid ">
                 <div className="flex items-center justify-between">
                   <label className="text-sm" htmlFor={label}>
                     {label}:
@@ -251,7 +251,7 @@ export const MetadataInputs = ({
   )
 }
 
-const emailInput: MetadataInput = {
+export const emailInput: MetadataInput = {
   type: 'email',
   name: 'email',
   label: 'Email',