+++ unlock-app/src/pages/_app.tsx
@@ -1,3 +1,4 @@
+import { SpeedInsights } from '@vercel/speed-insights/next'
 import React, { useEffect } from 'react'
 import type { AppProps } from 'next/app'
 import TagManager from 'react-gtm-module'
@@ -12,6 +13,15 @@ import { queryClient } from '~/config/queryClient'
 import { SessionProvider } from '~/hooks/useSession'
 import { ConnectModalProvider } from '~/hooks/useConnectModal'
 
+import { Inter } from 'next/font/google'
+
+const inter = Inter({
+  subsets: ['latin'],
+  style: ['normal'],
+  display: 'swap',
+  weight: ['400', '500', '600', '700'],
+})
+
 const UnlockApp = ({ Component, pageProps }: AppProps) => {
   useEffect(() => {
     if (!config.isServer) {
@@ -22,18 +32,21 @@ const UnlockApp = ({ Component, pageProps }: AppProps) => {
   }, [])
 
   return (
-    <QueryClientProvider client={queryClient}>
-      <SessionProvider>
-        <ConnectModalProvider>
-          <GlobalWrapper>
-            <ErrorBoundary fallback={(props) => <ErrorFallback {...props} />}>
-              <Component pageProps={pageProps} />
-            </ErrorBoundary>
-            <Toaster />
-          </GlobalWrapper>
-        </ConnectModalProvider>
-      </SessionProvider>
-    </QueryClientProvider>
+    <div className={inter.className}>
+      <QueryClientProvider client={queryClient}>
+        <SessionProvider>
+          <ConnectModalProvider>
+            <GlobalWrapper>
+              <ErrorBoundary fallback={(props) => <ErrorFallback {...props} />}>
+                <Component pageProps={pageProps} />
+              </ErrorBoundary>
+              <Toaster />
+            </GlobalWrapper>
+          </ConnectModalProvider>
+        </SessionProvider>
+      </QueryClientProvider>
+      <SpeedInsights />
+    </div>
   )
 }