+++ packages/unlock-js/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changes
 
+# 0.43.1
+
+- adding the ability to add a filter for approval when listing keys.
+
+# 0.43.0
+
+- allowing `maxNumberOfKeys` to be 0 when deploying locks.
+
 # 0.42.0
 
 - added support for `UnlockV13` and `PublicLockV14`